refactor(website): questionnaire-db.ts aufteilen (1227→3+450+299+301+121 Zeilen) [T001108]
