+++ website/src/pages/admin/fragebogen/[assignmentId].astro
@@ -6,7 +6,7 @@ import {
   listQAnswerOptionsForTemplate, listQAnswers, getQTemplate,
   listEvidenceByAssignment,
 } from '../../../lib/questionnaire-db';
-import { getDisplayScores } from '../../../lib/compute-scores';
+import { getDisplayScores } from '../../../lib/questionnaire-display';
 import { isSystemtestLoopEnabled } from '../../../lib/systemtest/feature-flag';
 import SystemtestReplayDrawer from '../../../components/SystemtestReplayDrawer.svelte';
