+++ website/src/pages/admin/projekte/[id].astro
@@ -15,7 +15,7 @@ import {
   listQAnswerOptionsForTemplate,
   listQAnswers,
 } from '../../../lib/questionnaire-db';
-import { getDisplayScores } from '../../../lib/compute-scores';
+import { getDisplayScores } from '../../../lib/questionnaire-display';
 import ProjectQuestionnairesPanel, { type AssignmentBundle } from '../../../components/admin/ProjectQuestionnairesPanel.astro';
 
 const session = await getSession(Astro.request.headers.get('cookie'));
