+++ website/src/pages/admin/tickets/[id].astro
@@ -10,7 +10,7 @@ import {
   listQAnswerOptionsForTemplate,
   listQAnswers,
 } from '../../../lib/questionnaire-db';
-import { getDisplayScores } from '../../../lib/compute-scores';
+import { getDisplayScores } from '../../../lib/questionnaire-display';
 import ProjectQuestionnairesPanel, { type AssignmentBundle } from '../../../components/admin/ProjectQuestionnairesPanel.astro';
 import GrillingStepper from '../../../components/admin/GrillingStepper.svelte';
 import { renderMarkdown } from '../../../lib/markdown';
